[LDBC]fix datatype convert for query11 result (#158)
fix datatype convert for query11 result

+++ ldbc_snb_interactive_implementations/nebula/src/main/java/org/ldbcouncil/snb/impls/workloads/nebula/NebulaDb.java
@@ -383,7 +383,7 @@ public class NebulaDb extends BaseDb<NebulaQueryStore>
             String personFirstName = record.get( 1 ).asString();
             String personLastName = record.get( 2 ).asString();
             String organizationName = record.get( 3 ).asString();
-            int organizationWorkFromYear = (int) record.get( 4 ).asLong();
+            int organizationWorkFromYear = record.get( 4 ).asInt();
             return new LdbcQuery11Result(
                     personId,
                     personFirstName,
